+++ .idea/misc.xml
@@ -1,5 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.23270440251572327" />
+        <entry key="app/src/main/res/layout/content_main.xml" value="0.23270440251572327" />
+        <entry key="app/src/main/res/layout/content_recipe_result.xml" value="0.2525597269624573" />
+        <entry key="app/src/main/res/layout/select_dressing_item.xml" value="0.29438405797101447" />
+        <entry key="app/src/main/res/menu/menu_main.xml" value="0.384375" />
+        <entry key="app/src/main/res/menu/menu_recipe.xml" value="0.384375" />
+      </map>
+    </option>
+  </component>
   <component name="ExternalStorageConfigurationManager" enabled="true" />
   <component name="NullableNotNullManager">
     <option name="myDefaultNullable" value="android.support.annotation.Nullable" />
