+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
             android:name="com.google.android.gms.ads.APPLICATION_ID"
             android:value="@string/ad_app_id"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:screenOrientation="portrait"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar"
@@ -32,7 +32,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".CreateRecipeActivity"
+            android:name=".ui.CreateRecipeActivity"
             android:screenOrientation="portrait"
             android:label="レシピ作成"
             android:theme="@style/AppTheme.NoActionBar" />
@@ -42,10 +42,10 @@
             android:label="レシピ詳細"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".TutorialActivity"
+            android:name=".ui.TutorialActivity"
             android:theme="@style/WelcomeScreenTheme" />
         <activity
-            android:name=".EditRecipeActivity"
+            android:name=".ui.EditRecipeActivity"
             android:screenOrientation="portrait"
             android:label="レシピ編集"
             android:theme="@style/AppTheme.NoActionBar" />
