+++ app/src/main/java/com/hammer/app/subwaysimulator/ui/CreateRecipeActivity.kt
@@ -1,4 +1,4 @@
-package com.hammer.app.subwaysimulator
+package com.hammer.app.subwaysimulator.ui
 
 import android.os.Bundle
 import android.app.AlertDialog
@@ -6,6 +6,14 @@ import android.view.View
 import android.widget.RadioButton
 import android.widget.RadioGroup
 import android.widget.Toast
+import com.hammer.app.subwaysimulator.model.Breads
+import com.hammer.app.subwaysimulator.model.Dressings
+import com.hammer.app.subwaysimulator.R
+import com.hammer.app.subwaysimulator.model.Recipe
+import com.hammer.app.subwaysimulator.model.amounts
+import com.hammer.app.subwaysimulator.model.breads
+import com.hammer.app.subwaysimulator.model.dressings
+import com.hammer.app.subwaysimulator.model.sandwiches
 import kotlinx.android.synthetic.main.create_recipe.*
 import kotlinx.android.synthetic.main.select_dressing_item.*
 import java.text.SimpleDateFormat
