+++ app/src/main/java/com/hammer/app/subwaysimulator/ui/TutorialActivity.kt
@@ -1,8 +1,9 @@
-package com.hammer.app.subwaysimulator
+package com.hammer.app.subwaysimulator.ui
 
 import android.app.Activity
 import android.os.Bundle
 import androidx.core.content.ContextCompat
+import com.hammer.app.subwaysimulator.R
 import com.stephentuso.welcome.*
 
 class TutorialActivity : WelcomeActivity() {
@@ -31,19 +32,19 @@ class TutorialActivity : WelcomeActivity() {
                 .defaultBackgroundColor(BackgroundColor(ContextCompat.getColor(this, R.color.colorPrimary)))
                 .page(TitlePage(R.drawable.icon, "SUBWAYシミュレーターへ"+ "\n" +"ようこそ！"))
                 .page(BasicPage(
-                        R.drawable.capture1,
+                    R.drawable.capture1,
                         "お気に入りを忘れないように", "このアプリではSUBWAYのサンドウィッチ・レシピを作成・保存することができます")
                         .background(BackgroundColor(ContextCompat.getColor(this, R.color.colorPrimary))))
                 .page(BasicPage(
-                        R.drawable.capture2,
+                    R.drawable.capture2,
                         "あなただけのレシピを作りましょう","SUBWAYの裏技完全対応！タップするだけでオリジナルレシピが作れます")
                         .background(BackgroundColor(ContextCompat.getColor(this, R.color.colorPrimary))))
                 .page(BasicPage(
-                        R.drawable.capture3,
+                    R.drawable.capture3,
                         "レシピを作ったら","作ったレシピは編集したり、画像として保存や友達へ共有もできます")
-                        .background(BackgroundColor(ContextCompat.getColor(this,R.color.colorPrimary))))
+                        .background(BackgroundColor(ContextCompat.getColor(this, R.color.colorPrimary))))
                 .page(BasicPage(
-                        R.drawable.capture4,
+                    R.drawable.capture4,
                         "レシピを削除したい場合は","レシピリストを長押しすることで削除できます")
                         .background(BackgroundColor(ContextCompat.getColor(this, R.color.colorPrimary))))
                 .swipeToDismiss(true)
